+++ include/game/item.h
@@ -15,6 +15,8 @@ public:
     virtual void update() = 0;
     virtual void render() = 0;
     virtual void reset() = 0;
+
+    void setUsed();
 };